backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
pandas==2.1.1
openpyxl==3.1.2
Werkzeug==2.3.7
